LayerProps as separate typings
