+++ package.json
@@ -3,7 +3,6 @@
   "description": "Flow map drawing layer for deck.gl",
   "author": "Teralytics AG",
   "license": "Apache-2.0",
-  "version": "6.1.0",
   "private": true,
   "keywords": [
     "flows",
