+++ packages/core/src/AnimatedFlowLinesLayer/AnimatedFlowLinesLayer.ts
@@ -46,7 +46,7 @@ import { RGBA } from '../colors';
 import { AccessorObjectInfo, Flow } from '../types';
 import FragmentShader from './AnimatedFlowLinesLayerFragment.glsl';
 import VertexShader from './AnimatedFlowLinesLayerVertex.glsl';
-import {LayerProps} from "@deck.gl/core/lib/layer";
+import {LayerProps} from "../LayerProps";
 
 export interface Props extends LayerProps {
   id: string;
